% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{e_append1_p}
\alias{e_append1_p}
\alias{e_append2_p}
\title{Append data}
\usage{
e_append1_p(proxy, series.index = NULL, data, x, y)

e_append2_p(proxy, series.index = NULL, data, x, y, z)
}
\arguments{
\item{proxy}{An echarts4r proxy as returned by \code{\link{echarts4rProxy}}.}

\item{series.index}{Index of serie to append to (starts from 0).}

\item{data}{Data.frame containing data to append.}

\item{x, y, z}{Columns names to plot.}
}
\description{
Append data.
}
\details{
Currently not all types of series supported incremental rendering when using appendData. 
Only these types of series support it: \code{\link{e_scatter}} and \code{\link{e_line}} of pure echarts, and 
\code{\link{e_scatter_3d}}, and \code{\link{e_line_3d}} of echarts-gl.
}
\examples{
if(interactive()){
  library(shiny)
  
  ui <- fluidPage(
    actionButton("add", "Add Data"),
    echarts4rOutput("plot")
  )
  
  server <- function(input, output, session){
  
    data <- data.frame(x = rnorm(10, 5, 3), y = rnorm(10, 50, 12))
    
    react <- eventReactive(input$add, {
      set.seed(sample(1:1000, 1))
      data.frame(x = rnorm(10, 5, 2), y = rnorm(10, 50, 10))
    })
    
    output$plot <- renderEcharts4r({
      data \%>\% 
       e_charts(x) \%>\% 
       e_scatter(y)
    })
    
    observeEvent(input$add, {
      echarts4rProxy("plot") \%>\% 
        e_append1_p(0, react(), x, y)
    })
    
  }
  
  shinyApp(ui, server)
  
}

}
