% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{band}
\alias{band}
\alias{e_band}
\alias{e_band_}
\title{Confidence bands}
\usage{
e_band(e, serie, min, max, stack = "confidence-band", ...)

e_band_(e, serie, min, max, stack = "confidence-band", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie, min, max}{series.}

\item{stack}{Name of stack.}

\item{...}{All options must be of vectors or lists of length 2 where the first argument is 
for the lower bound and the second for the upper bound, see examples.}
}
\description{
Add confidence bands
}
\examples{
df <- data.frame(
  x = 1:20,
  y = runif(20, 10 , 30),
  lwr = runif(20, 5, 10),
  upr = runif(20, 30, 40)
)

df \%>\% 
  e_charts(x) \%>\% 
  e_line(y) \%>\% 
  e_band(y, lwr, upr)

}
