% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_bar}
\alias{e_bar}
\alias{e_line}
\title{Bar & Line chart}
\usage{
e_bar(e, serie, name = NULL, ...)

e_line(e, serie, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4} object as returned by \code{e_charts}.}

\item{serie}{Column name of serie to plot on y axis.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add bar or line serie.
}
\examples{
USArrests \%>\% 
  dplyr::mutate(
    state = row.names(.)
  ) \%>\% 
  e_charts(state) \%>\% 
  e_bar(Murder, name = "Euwww") \%>\% 
  e_line(Rape)

}
