% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildings_sample_json}
\alias{buildings_sample_json}
\title{Buildings sample set in json format}
\format{
A sample list of buildings
}
\source{
echarts
\url{https://echarts.apache.org/examples/data-gl/asset/data/buildings.json}
}
\usage{
buildings_sample_json
}
\description{
Buildings sample set in json format
}
\keyword{datasets}
