% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.build_sun}
\alias{.build_sun}
\title{Adding styles to hierarchical data by item names or levels}
\usage{
.build_sun(e, styles = NULL, names = NULL, levels = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{styles}{Style lists, expects a \code{vector}, defaults to \code{NULL}.}

\item{names}{Names of items to style, expects a \code{list}, defaults to \code{NULL}.}

\item{levels}{Hierarchical levels to style, expects a \code{list}, defaults to \code{NULL}.}
}
\value{
updated hierarchy in json list format
recursive json-list traversal, append style on matching level and/or name
}
\description{
Adding styles to hierarchical data by item names or levels
}
\author{
helgasoft.com
}
