% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{e_area}
\alias{e_area}
\alias{e_area_}
\title{Area}
\usage{
e_area(e, serie, bind, name = NULL, legend = TRUE, y.index = 0,
  x.index = 0, ...)

e_area_(e, serie, bind = NULL, name = NULL, legend = TRUE,
  y.index = 0, x.index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{bind}{Binding between datasets, namely for use of \code{\link{e_brush}}.}

\item{name}{name of the serie.}

\item{legend}{Whether to add serie to legend.}

\item{y.index}{Indexes of x and y axis.}

\item{x.index}{Indexes of x and y axis.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Add area serie.
}
\examples{
CO2 \%>\% 
  group_by(Plant) \%>\% 
  e_charts(conc) \%>\% 
  e_area(uptake)

}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#series-line}
}
