% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_axis_labels}
\alias{e_axis_labels}
\title{Axis Labels}
\usage{
e_axis_labels(e, x = "", y = "")
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{x, y}{Labels of axes.}
}
\description{
Convenience function to add axis labels.
}
\examples{
cars \%>\%
  e_charts(speed) \%>\%
  e_scatter(dist) \%>\%
  e_axis_labels(
    x = "speed",
    y = "distance"
  )
}
