% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_axis_pointer}
\alias{e_axis_pointer}
\title{Axis pointer}
\usage{
e_axis_pointer(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise axis pointer.
}
