% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_axis_stagger}
\alias{e_axis_stagger}
\title{Stagger Axis Labels}
\usage{
e_axis_stagger(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}
}
\description{
Stagger axis labels.
}
\examples{
df <- data.frame(
 x = c("a very long label", "Another long label"),
 y = 1:2
)

df \%>\% 
 e_charts(x, width = 150) \%>\% 
 e_bar(y) \%>\% 
 e_axis_stagger()

}
