% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{e_bar}
\alias{e_bar}
\alias{e_bar_}
\title{Bar and Line chart}
\usage{
e_bar(e, serie, bind, name = NULL, legend = TRUE, y.index = 0,
  x.index = 0, ...)

e_bar_(e, serie, bind = NULL, name = NULL, legend = TRUE,
  y.index = 0, x.index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{bind}{Binding between datasets, namely for use of \code{\link{e_brush}}.}

\item{name}{name of the serie.}

\item{legend}{Whether to add serie to legend.}

\item{x.index, y.index}{Indexes of x and y axis.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Add bar serie.
}
\examples{
USArrests \%>\% 
  dplyr::mutate(
    State = row.names(.),
    Rape = -Rape
  ) \%>\% 
  e_charts(State) \%>\% 
  e_bar(Murder) \%>\% 
  e_bar(Rape, name = "Sick basterd", x.index = 1) # secondary x axis

}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#series-bar}
}
