% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_bar_3d}
\alias{e_bar_3d}
\title{Bars 3D}
\usage{
e_bar_3d(e, x, y, z, coord.system = "cartesian3D", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{coord.system}{Coordinate system to use, one of \code{cartesian3D}, \code{geo3D}, \code{globe}.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add 3D bars
}
\examples{
url <- paste0("https://ecomfe.github.io/echarts-examples/",
              "public/data-gl/asset/data/population.json")
data <- jsonlite::fromJSON(url)
data <- as.data.frame(data)
names(data) <- c("lon", "lat", "value")

data \%>\% 
  e_charts() \%>\% 
  e_globe(
    environment = e_stars_texture(),
    base.texture = e_globe_texture()
  ) \%>\% 
  e_bar_3d(lon, lat, value, "globe") \%>\% 
  e_visual_map()

}
