% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{e_barRange}
\alias{e_barRange}
\title{Bar range chart}
\usage{
e_barRange(
  e,
  serie_min,
  serie_max,
  barWidth = 10,
  borderRadius = 5,
  margin = 10,
  textSymbol = "°F",
  legend = TRUE,
  name = "barRange",
  ...
)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{serie_min}{Column name of minimum value in the serie.}

\item{serie_max}{Column name of maximum value in the serie.}

\item{barWidth}{width of each bar}

\item{borderRadius}{roundness of the bar ends}

\item{margin}{space between the text labels and the bar}

\item{textSymbol}{string for the label to end with}

\item{legend}{Whether to add serie to legend.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Draw a bar range plot with labels on each end.
}
\examples{

df <- iris |>
  dplyr::group_by(Species) |>
  dplyr::summarise(min_length = min(Sepal.Length),
                   max_length = max(Sepal.Length))

df |> e_chart(Species) |>
  e_barRange(serie_min="min_length",
             serie_max="max_length",
             textSymbol = '"'
  )
}
\seealso{
\href{https://github.com/apache/echarts-custom-series/tree/main/custom-series/barRange}{official documentation}
}
