% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_boxplot}
\alias{e_boxplot}
\title{Boxplot}
\usage{
e_boxplot(e, serie, name = NULL, outliers = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4} object as returned by \code{e_charts}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw boxplot.
}
\examples{
df <- data.frame(
  x = rnorm(10),
  y = rnorm(10, 2, 1)
)

df \%>\% 
  e_charts() \%>\% 
  e_boxplot(y) \%>\% 
  e_boxplot(x)

}
