% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{e_button}
\alias{e_button}
\title{Button}
\usage{
e_button(e, event, id, ..., position = "top",
  tag = htmltools::tags$button)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{event}{Name of event to trigger.}

\item{id}{A valid CSS id.}

\item{...}{Content of the button, complient with \code{htmltools}.}

\item{position}{Position of button, \code{top} or \code{bottom}.}

\item{tag}{A Valid \link[htmltools]{tags} function.}
}
\description{
Add a button to your visualisation.
}
\section{Events}{

\itemize{
  \item{\code{highlight}}
  \item{\code{downplay}}
  \item{\code{showtip}}
  \item{\code{hidetip}}
  \item{\code{zoom}}
  \item{\code{visual_map_range}}
  \item{\code{pie_select}}
  \item{\code{pie_unselect}}
}
}

\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_charts(Sepal.Length) \%>\% 
  e_line(Sepal.Width) \%>\% 
  e_line(Petal.Length) \%>\% 
  e_highlight(series.name = "setosa", btn = "myBtn") \%>\% 
  e_button("highlight", "myBtn", "highlight stuff")

}
