% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{e_calendar}
\alias{e_calendar}
\title{Calendar}
\usage{
e_calendar(e, range, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Calendar
}
