% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts4.R
\name{e_charts}
\alias{e_charts}
\title{Initialise}
\usage{
e_charts(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}
}
\description{
Initialise a chart.
}
