% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{e_color_range}
\alias{e_color_range}
\alias{e_color_range_}
\title{Color range}
\usage{
e_color_range(data, input, output, colors = c("#bf444c", "#d88273",
  "#f6efa6"), ...)

e_color_range_(data, input, output, colors = c("#bf444c", "#d88273",
  "#f6efa6"), ...)
}
\arguments{
\item{data}{Data.frame in which to find column names.}

\item{input, output}{Input and output columns.}

\item{colors}{Colors to pass to \code{\link{colorRampPalette}}.}

\item{...}{Any other argument to pass to \code{\link{colorRampPalette}}.}
}
\description{
Build manual color range
}
\examples{
df <- data.frame(val = 1:10)

e_color_range(df, val, colors)
  
}
