% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{e_common}
\alias{e_common}
\title{General options}
\usage{
e_common(font_family = NULL, theme = NULL)
}
\arguments{
\item{font_family}{Font family.}

\item{theme}{A theme.}
}
\description{
General options
}
