% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_connect}
\alias{e_connect}
\title{Set Group}
\usage{
e_connect(e, id)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{id}{Id of chart to connect with.}
}
\description{
Set group to connect charts together.
}
\examples{
e1 <- cars \%>\% 
  e_charts(
    speed,
    elementId = "chart1"
  ) \%>\% 
  e_scatter(dist)
  
e2 <- cars \%>\% 
  e_charts(
    dist,
    elementId = "chart2"
  ) \%>\% 
  e_scatter(speed) \%>\% 
  e_connect("chart1") # connect

htmltools::browsable(
  htmltools::div(
    class = "container",
      htmltools::div(
        class = "col-md-6",
        e1
      ),
      htmltools::div(
        class = "col-md-6",
        e2
      )
   )
)

}
