% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_correlations}
\alias{e_correlations}
\title{Correlation}
\usage{
e_correlations(e, order = NULL, visual_map = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{order}{Ordering method, passed to \link[corrplot]{corrMatOrder}.}

\item{visual_map}{Whether to add the visual map.}

\item{...}{Any argument to pass to \code{\link{e_heatmap}} and \code{\link{e_visual_map}}.}
}
\description{
Correlation
}
\examples{
cor(mtcars) \%>\% 
  e_charts() \%>\% 
  e_correlations(
    order = "hclust",
    visual_map = FALSE
  ) \%>\% 
  e_visual_map(
   min = -1, 
   max = 1
  )

}
