% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_dims}
\alias{e_dims}
\title{Dimensions}
\usage{
e_dims(e, height = "auto", width = "auto")
}
\description{
Sets the dimensions of the chart _internally._
This will only affect the dimensions of the
chart within its parent container.
Use the `height` and `width` arguments of 
[e_charts] if you want to change the dimensions
of said parent (recommended).
}
