% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{e_dispatch_action_p}
\alias{e_dispatch_action_p}
\title{Dispatch Action}
\usage{
e_dispatch_action_p(proxy, type, ...)
}
\arguments{
\item{proxy}{An echarts4r proxy as returned by \code{\link{echarts4rProxy}}.}

\item{type}{Type of action to dispatch, i.e.: \code{highlight}.}

\item{...}{Named options.}
}
\description{
Create your own proxies, essentially a wrapper around the 
\href{https://ecomfe.github.io/echarts-doc/public/en/api.html#action}{action API}.
}
