% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{e_doughnut}
\alias{e_doughnut}
\alias{e_doughnut_}
\title{Segmented Doughnut}
\usage{
e_doughnut(e, serie, color, rm_x = TRUE, rm_y = TRUE, ...)

e_doughnut_(e, serie, color = NULL, rm_x = TRUE, rm_y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{serie}{Column name of serie to plot.}

\item{color}{Color to plot.}

\item{rm_x, rm_y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Draw segmented doughnut.
}
\examples{
df <- data.frame(val = c(0.6, 0.5, 0.4))

df |>
  e_charts() |>
  e_liquid(val) |>
  e_theme("dark")
}
\seealso{
\href{https://github.com/apache/echarts-custom-series/tree/main/custom-series/segmentedDoughnut}{official documentation}
}
