% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_flip_coords}
\alias{e_flip_coords}
\title{Flip coordinates}
\usage{
e_flip_coords(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}
}
\description{
Flip cartesian 2D coordinates.
}
\examples{
df <- data.frame(
  x = LETTERS[1:5],
  y = runif(5, 1, 5),
  z = runif(5, 3, 10)
)

df \%>\%
 e_charts(x) \%>\%
 e_bar(y) \%>\% 
 e_line(z) -> plot
 
plot # normal
e_flip_coords(plot) # flip

}
