% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_flow_gl}
\alias{e_flow_gl}
\title{Flow GL}
\usage{
e_flow_gl(e, y, sx, sy, color, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{y}{Vector position on the y axis.}

\item{sx, sy}{Velocity in respective axis.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}

\item{Color}{Vector color.}
}
\description{
Flow GL
}
\examples{
vectors <- expand.grid(0:9, 0:9)
names(vectors) <- c("x", "y")
vectors$sx <- rnorm(100)
vectors$sy <- rnorm(100)
vectors$color <- log10(runif(100, 1, 10))

vectors \%>\% 
  e_charts(x) \%>\% 
  e_flow_gl(y, sx, sy, color) \%>\% 
  e_visual_map(
    min = 0, max = 1, # log 10
    dimension = 4, # x = 0, y = 1, sx = 3, sy = 4
    show = FALSE, # hide
    inRange = list(
      color = c('#313695', '#4575b4', '#74add1', '#abd9e9', '#e0f3f8',
                '#ffffbf', '#fee090', '#fdae61', '#f46d43', '#d73027', '#a50026')
    )
  ) \%>\% 
  e_x_axis(
    splitLine = list(show = FALSE)
  ) \%>\% 
  e_y_axis(
    splitLine = list(show = FALSE)
  )   

}
