% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_funnel}
\alias{e_funnel}
\title{Funnel}
\usage{
e_funnel(e, values, labels, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{values, labels}{Values and labels of funnel.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add a funnel.
}
\examples{
funnel <- data.frame(stage = c("View", "Click", "Purchase"), value = c(80, 30, 20))

funnel \%>\% 
  e_charts() \%>\% 
  e_funnel(value, stage)

}
