% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_gauge}
\alias{e_gauge}
\title{Gauge}
\usage{
e_gauge(e, value, name, rm.x = TRUE, rm.y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{value}{Value to gauge.}

\item{name}{Text on gauge.}

\item{rm.x, rm.y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Plot a gauge.
}
\examples{
e_charts() \%>\% 
  e_gauge(57, "PERCENT")

}
