% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{e_geo_3d}
\alias{e_geo_3d}
\title{Geo}
\usage{
e_geo_3d(e, serie, color, type = "world", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{color}{Color.}

\item{type}{Map type.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Create a 3d geo plot.
}
\examples{
choropleth <- data.frame(
  countries = c("France", "Brazil", "China", "Russia", "Canada", "India", "United States",
                "Argentina", "Australia"),
  height = runif(9, 1, 5),
  color = c("#F7FBFF", "#DEEBF7", "#C6DBEF", "#9ECAE1", "#6BAED6", "#4292C6", 
            "#2171B5", "#08519C", "#08306B")
)

choropleth \%>\% 
  e_charts(countries) \%>\% 
  e_geo_3d(height, color)

}
