% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{e_get_data}
\alias{e_get_data}
\title{Get data}
\usage{
e_get_data(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}
}
\description{
Get data passed to \code{\link{e_charts}}.
}
