% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{e_get_zr}
\alias{e_get_zr}
\title{Blank Area}
\usage{
e_get_zr()
}
\description{
Use this function to capture a click on a
blank area of the canvas.
Note that this may stops other "click" events
from working.
}
