% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe.R
\name{e_globe_bar}
\alias{e_globe_bar}
\title{Globe}
\usage{
e_globe_bar(e, lon, lat, value, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{lon, lat}{Coordinates.}

\item{value}{Bar height.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Globe
}
\examples{
url <- paste0("https://ecomfe.github.io/echarts-examples/",
              "public/data-gl/asset/data/population.json")
data <- jsonlite::fromJSON(url)
data <- as.data.frame(data)
names(data) <- c("lon", "lat", "value")

data \%>\% 
  e_charts() \%>\% 
  e_globe_bar(lon, lat, value)

}
