% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3d.R
\name{e_grid_3d}
\alias{e_grid_3d}
\title{Grid}
\usage{
e_grid_3d(e, index = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{index}{Index of axis to customise.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise grid.
}
\examples{
USArrests \%>\% 
  e_charts(UrbanPop) \%>\% 
  e_line(Assault, smooth = TRUE) \%>\% 
  e_area(Murder, y.index = 1, x.index = 1) \%>\% 
  e_y_axis(gridIndex = 1) \%>\%
  e_x_axis(gridIndex = 1) \%>\% 
  e_grid(height = "40\%") \%>\% 
  e_grid(height = "40\%", top = "55\%")

}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#grid3D}
}
