% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_heatmap}
\alias{e_heatmap}
\title{Heatmap}
\usage{
e_heatmap(e, x, y, z, name = NULL, ...)
}
\description{
Heatmap
}
\examples{
v <- c(1:10)
matrix <- data.frame(
  x = sample(v, 300, replace = TRUE), 
  y = sample(v, 300, replace = TRUE), 
  z = rnorm(300, 10, 1)
) \%>\% 
  dplyr::group_by(x, y) \%>\% 
  dplyr::summarise(z = sum(z)) \%>\% 
  dplyr::ungroup()

matrix \%>\% 
  e_charts() \%>\% 
  e_heatmap(x, y, z)

}
