% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_hide_grid_lines}
\alias{e_hide_grid_lines}
\title{' Hide Grid Lines}
\usage{
e_hide_grid_lines(e, which = c("x", "y"))
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{which}{Which axis grid lines to hide.}
}
\description{
A convenience function to easily hide grid lines.
}
\examples{
cars \%>\%
 e_charts(speed) \%>\% 
 e_scatter(dist) \%>\% 
 e_hide_grid_lines()
 
}
