% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{e_inspect}
\alias{e_inspect}
\title{Inspect}
\usage{
e_inspect(e, json = FALSE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{json}{Whether to return the JSON, otherwise returns a \code{list}.}

\item{...}{Additional options to pass to \link[jsonlite]{toJSON}.}
}
\value{
Returns a \code{list} if \code{json} is \code{FALSE}.
}
\description{
Inspect JSON options built by the package.
}
\note{
Must be passed as last option.
}
\examples{
mtcars \%>\% 
  e_charts(mpg) \%>\% 
  e_scatter(wt, qsec) \%>\% 
  e_inspect(
    json = TRUE,
    pretty = TRUE
  )

}
