% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_keras_history}
\alias{e_keras_history}
\title{History}
\usage{
e_keras_history(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}
}
\description{
Plot \link[keras] history in R.
}
\examples{
\dontrun{
history <- model \%>\% 
 fit(...)

history \%>\%
 e_charts() \%>\%
 e_keras_history()
}

}
