% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_legend}
\alias{e_legend}
\title{Legend}
\usage{
e_legend(e, show = TRUE, type = c("plain", "scroll"), ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{show}{Set to \code{FALSE} to hide the legend.}

\item{type}{Type of legend, \code{plain} or \code{scroll}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Customise the legend.
}
\examples{
mtcars \%>\% 
  head() \%>\% 
  dplyr::mutate(model = row.names(.)) \%>\% 
  e_charts(model) \%>\% 
  e_pie(carb) \%>\% 
  e_legend(FALSE)

}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#legend}{Additional arguments}
}
