% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_line_3d}
\alias{e_line_3d}
\title{Lines 3D}
\usage{
e_line_3d(e, source.lon, source.lat, target.lon, target.lat, name = NULL,
  coord.system = "globe", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{source.lon, source.lat, target.lon, target.lat}{coordinates.}

\item{name}{name of the serie.}

\item{coord.system}{Coordinate system to use, one of \code{cartesian3D}, \code{globe}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add 3D lines.
}
\examples{
flights <- read.csv(
  paste0("https://raw.githubusercontent.com/plotly/datasets/",
         "master/2011_february_aa_flight_paths.csv")
)

flights \%>\% 
  e_charts() \%>\% 
  e_globe(
    base.texture = e_map_texture(),
    height.texture = e_map_texture(),
    environment = e_stars_texture(),
    displacementScale = 0.05
  ) \%>\% 
  e_line_3d(
    start_lon, 
    start_lat, 
    end_lon, 
    end_lat,
    name = "flights"
  )
  
flights \%>\% 
  e_charts() \%>\% 
  e_geo_3d() \%>\% 
  e_line_3d(
    start_lon, 
    start_lat, 
    end_lon, 
    end_lat,
    coord.system = "geo3D"
  )

}
