% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{e_liquid}
\alias{e_liquid}
\alias{e_liquid_}
\title{Liquid fill}
\usage{
e_liquid(e, serie, color, rm_x = TRUE, rm_y = TRUE, ...)

e_liquid_(e, serie, color = NULL, rm_x = TRUE, rm_y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{color}{Color to plot.}

\item{rm_x, rm_y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Draw liquid fill.
}
\examples{
df <- data.frame(val = c(0.6, 0.5, 0.4))

df \%>\% 
  e_charts() \%>\% 
  e_liquid(val) \%>\% 
  e_theme("dark")

}
\seealso{
\href{https://github.com/ecomfe/echarts-liquidfill}{official documentation}
}
