% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_lm}
\alias{e_lm}
\title{Linear Regression}
\usage{
e_lm(e, formula, name = NULL, legend = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{formula}{formula to pass to \code{\link{lm}}.}

\item{name, }{series name.}

\item{legend}{Whether to plot legend.}

\item{...}{Additional arguments to pass to \code{\link{e_line}}.}
}
\description{
Plot fitted.
}
\examples{
mtcars \%>\% 
  e_charts(mpg) \%>\% 
  e_scatter(qsec) \%>\% 
  e_lm(qsec ~ mpg, name = "y = ax + b")
  
CO2 \%>\% 
  e_charts(conc) \%>\% 
  e_scatter(uptake) \%>\% 
  e_lm(uptake ~ conc, name = "regression")

}
