% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{e_merge}
\alias{e_merge}
\title{Merge options in chart, used in e_mark}
\usage{
e_merge(proxy)
}
\arguments{
\item{proxy}{An echarts4r proxy as returned by \code{\link{echarts4rProxy}}.}
}
\description{
Merge options in chart, used in e_mark
}
