% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_parallel}
\alias{e_parallel}
\title{Parallel}
\usage{
e_parallel(e, ..., name = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}

\item{name}{name of the serie.}
}
\description{
Draw parallel coordinates.
}
\examples{
df <- data.frame(
  price = rnorm(5, 10),
  amount = rnorm(5, 15),
  letter = LETTERS[1:5]
)

df \%>\% 
  e_charts() \%>\% 
  e_parallel(price, amount, letter)

}
