% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_pie}
\alias{e_pie}
\title{Pie}
\usage{
e_pie(e, serie, label, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{label}{Labels of slices.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw pie and donut charts.
}
\examples{
mtcars \%>\% 
  head() \%>\% 
  dplyr::mutate(model = row.names(.)) \%>\% 
  e_charts() \%>\% 
  e_pie(carb, model)

}
