% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_polar}
\alias{e_polar}
\title{Polar}
\usage{
e_polar(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise polar coordinates.
}
