% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_radar}
\alias{e_radar}
\title{Radar}
\usage{
e_radar(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Customise radar.
}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#radar}
}
