% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_radar_opts}
\alias{e_radar_opts}
\title{Radar axis}
\usage{
e_radar_opts(e, index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{index}{Index of axis to customise.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Radar axis setup and options.
}
