% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_radius_axis}
\alias{e_radius_axis}
\title{Radius axis}
\usage{
e_radius_axis(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise radius axis.
}
