% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_scatter}
\alias{e_scatter}
\title{Scatter}
\usage{
e_scatter(e, serie, size, scale = "* 1", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{size}{Column name containing size of points.}

\item{scale}{Scale for \code{size}, defaults to \code{* 1} which multiplies the size 
by \code{1} (equivalent to no multiplier).}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add a scatter serie.
}
\examples{
USArrests \%>\% 
  e_charts(Assault) \%>\% 
  e_scatter(Murder, Rape)

}
