% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_scatter_3d}
\alias{e_scatter_3d}
\title{Scatter 3D}
\usage{
e_scatter_3d(e, y, z, color, size, coord.system = "cartesian3D",
  name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{y, z}{Coordinates.}

\item{coord.system}{Coordinate system to use, one of \code{geo3D}, \code{globe}, or \code{cartesian3D}.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add 3D scatter.
}
\examples{
v <- LETTERS[1:10]
matrix <- data.frame(
  x = sample(v, 300, replace = TRUE), 
  y = sample(v, 300, replace = TRUE), 
  z = rnorm(300, 10, 1),
  stringsAsFactors = FALSE
) \%>\% 
  dplyr::group_by(x, y) \%>\% 
  dplyr::summarise(
    z = sum(z), 
    size = sum(size),
    color = sum(color)
  ) \%>\% 
  dplyr::ungroup() 
  
matrix \%>\% 
  e_charts(x) \%>\% 
  e_scatter_3d(y, z) \%>\% 
  e_visual_map()
  
airports <- read.csv(
  paste0("https://raw.githubusercontent.com/plotly/datasets/",
         "master/2011_february_us_airport_traffic.csv")
)

airports \%>\% 
  e_charts(long) \%>\% 
  e_globe(
    environment = e_stars_texture(),
    base.texture = e_globe_texture(), 
    globeOuterRadius = 100
  ) \%>\% 
  e_scatter_3d(lat, cnt, coord.system = "globe", blendMode = 'lighter') \%>\% 
  e_visual_map()

}
