% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_scatter_gl}
\alias{e_scatter_gl}
\title{Scatter GL}
\usage{
e_scatter_gl(e, y, z, name = NULL, coord.system = "geo", rm.x = TRUE,
  rm.y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{y, z}{Column names containing y and z data.}

\item{name}{name of the serie.}

\item{coord.system}{Coordinate system to plot against.}

\item{rm.x, rm.y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw scatter GL.
}
\examples{
quakes \%>\% 
  e_charts(long) \%>\% 
  e_geo(
    roam = TRUE,
    boundingCoords = list(
      c(185, - 10),
      c(165, -40)
     )
  ) \%>\% 
  e_scatter_gl(lat, depth)

}
\seealso{
\href{Additional arguments}{http://echarts.baidu.com/option-gl.html#series-scatterGL}
}
