% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webGL.R
\name{e_scatterGL}
\alias{e_scatterGL}
\title{Scatter GL}
\usage{
e_scatterGL(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Scatter GL
}
