% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_single_axis}
\alias{e_single_axis}
\title{Single Axis}
\usage{
e_single_axis(e, index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{index}{Index of axis to customise.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Setup single axis.
}
\examples{
df <- data.frame(
  axis = LETTERS[1:10],
  value = runif(10, 3, 20),
  size = runif(10, 3, 20)
)

df \%>\%
  e_charts(axis) \%>\%
  e_single_axis() \%>\% # add the single axis
  e_scatter(
    value,
    size,
    coord_system = "singleAxis"
  )
}
