% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_step}
\alias{e_step}
\title{Step}
\usage{
e_step(e, serie, step = c("start", "middle", "end"), fill = FALSE,
  name = NULL, y.index = 0, x.index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add step serie.
}
\examples{
USArrests \%>\% 
  dplyr::mutate(State = row.names(.)) \%>\% 
  e_charts(State) \%>\%
  e_step(Murder, name = "Start", step = "start", fill = TRUE) \%>\% 
  e_step(Rape, name = "Middle", step = "middle") \%>\% 
  e_step(Assault, name = "End", step = "end") \%>\% 
  e_tooltip(trigger = "axis")

}
