% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_text_style}
\alias{e_text_style}
\title{Text style}
\usage{
e_text_style(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Define global font style.
}
\examples{
cars \%>\% 
  e_charts(dist) \%>\% 
  e_scatter(speed) \%>\% 
  e_text_style(
    color = "blue",
    fontStyle = "italic"
  )

}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#textStyle}{official documentation}
}
