% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{e_timeline}
\alias{e_timeline}
\title{Timeline}
\usage{
e_timeline(data, x, width = NULL, height = NULL, elementId = NULL,
  dispose = TRUE, renderer = "canvas", ...)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x}{Column name containing x axis.}

\item{width}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}

\item{dispose}{Set to \code{TRUE} to force redraw of chart, set to \code{FALSE} to update.}

\item{renderer}{Renderer, takes \code{canvas} (default) or \code{svg}.}

\item{...}{Options passed to \code{\link{e_timeline_serie_opts}}.}
}
\description{
Create a timeline.
}
\details{
Group your data by steps with \link[dplyr]{group_by}, see examples.
}
\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_timeline(Sepal.Length, autoPlay = TRUE) \%>\% 
  e_timeline_serie(Sepal.Width) \%>\% 
  e_timeline_serie(Petal.Length) 

}
