% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{e_timeline_base_opts}
\alias{e_timeline_base_opts}
\title{Base Options}
\usage{
e_timeline_base_opts(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_timeline}}.}

\item{...}{Any option.}
}
\description{
Add base options
}
