% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{e_timeline_opts}
\alias{e_timeline_opts}
\title{Timeline options}
\usage{
e_timeline_opts(e, axis_type = "category", data = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_timeline}}.}

\item{axis_type}{Type of axis, \code{category}, \code{value} or \code{time}.}

\item{data}{Timeline steps.}

\item{...}{Any other named option.}
}
\description{
Set timeline options.
}
\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_timeline(Sepal.Length) \%>\% 
  e_timeline_serie_("Sepal.Width") \%>\% 
  e_timeline_serie_("Petal.Length") \%>\% 
  e_timeline_opts(
    autoPlay = TRUE,
    rewind = TRUE
  )

}
