% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{e_timeline_serie_opts}
\alias{e_timeline_serie_opts}
\title{Timeline series options}
\usage{
e_timeline_serie_opts(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_timeline}}.}

\item{...}{Named options \emph{the same length as the number of steps} (groups).}
}
\description{
Timeline series options
}
\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_timeline(Sepal.Length, autoPlay = TRUE) \%>\% 
  e_timeline_serie(Sepal.Width) \%>\% 
  e_timeline_serie(Petal.Length) \%>\% 
  e_timeline_serie_opts(
    title = list(
      list(text = "setosa"),
      list(text = "versicolor"),
      list(text = "virginica")
    )
  )

}
