% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{e_title_matrix}
\alias{e_title_matrix}
\title{Generate Chart Titles for Matrix}
\usage{
e_title_matrix(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
helper function for creating titles for every plot in a geofacet style
matrix. This generates the title using the name of the series.
}
\seealso{
\href{https://echarts.apache.org/en/option.html#title}{Additional
  arguments}
}
