% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_toolbox_interface}
\alias{e_toolbox_interface}
\title{Toolbox}
\usage{
e_toolbox_interface(e, feature)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{feature}{Feature to add, defaults to all.}
}
\description{
Add a toolbox.
}
\examples{
USArrests \%>\% 
  e_charts(UrbanPop) \%>\% 
  e_line(Assault) \%>\% 
  e_area(Murder, y.index = 1, x.index = 1) \%>\% 
  e_datazoom(x.index = 0) 

}
