% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{e_toolbox_restore}
\alias{e_toolbox_restore}
\title{Restore Toolbox}
\usage{
e_toolbox_restore(e, btn = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{btn}{A \code{\link{e_button}} id.}
}
\description{
Restore Toolbox.
}
