% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_utc}
\alias{e_utc}
\title{Use UTC}
\usage{
e_utc(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}
}
\description{
Use UTC
}
