% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{e_violin}
\alias{e_violin}
\title{Violin chart}
\usage{
e_violin(
  e,
  data,
  x,
  y,
  show_scatter = TRUE,
  violin_color = NULL,
  violin_opacity = 0.8,
  scatter_color = NULL,
  scatter_size = 6,
  bin_count = 20,
  bandwidth_scale = 1.5,
  ...
)
}
\arguments{
\item{show_scatter}{TRUE/FALSE to show scatter dots}

\item{violin_color}{violin color}

\item{violin_opacity}{opacity of the violin, between 0-1.}

\item{scatter_color}{scatter color}

\item{scatter_size}{size of scatter dots}

\item{bin_count}{Count of bins the data is divided into when calculating the distribution.}

\item{bandwidth_scale}{smoothness of the violin's shape by adjusting the kernel density estimation (KDE) bandwidth. A higher value smoothes the shape.}

\item{...}{}
}
\description{
Draw a violin chart with scattered dots.
}
\examples{
e_chart() |> e_violin(data = iris, x =  "Species", y = "Sepal.Length") |> e_jitter(jitter = 100)

}
\seealso{
\href{https://github.com/apache/echarts-custom-series/tree/main/custom-series/segmentedDoughnut}{official documentation}
}
