% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{e_zoom}
\alias{e_zoom}
\title{Zoom}
\usage{
e_zoom(e, ..., btn = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any options, see \href{https://echarts.apache.org/en/api.html#action.dataZoom.dataZoom}{official documentation}}

\item{btn}{A \code{\link{e_button}} id.}
}
\description{
Zoom on a region.
}
\examples{
cars \%>\% 
  e_charts(dist) \%>\% 
  e_scatter(speed) \%>\% 
  e_datazoom() \%>\% 
  e_zoom(
    dataZoomIndex = 0,
    start = 20,
    end = 40,
    btn = "BUTTON"
  ) \%>\% 
  e_button("BUTTON", "Zoom in")

}
