% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts4r.R
\name{echarts4r-shiny}
\alias{echarts4r-shiny}
\alias{echarts4rOutput}
\alias{renderEcharts4r}
\alias{echarts4rProxy}
\title{Shiny bindings for echarts4r}
\usage{
echarts4rOutput(outputId, width = "100\%", height = "400px")

renderEcharts4r(expr, env = parent.frame(), quoted = FALSE)

echarts4rProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{output variable to read from.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a echarts4r}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{Target chart id.}

\item{session}{Shiny session.}
}
\description{
Output and render functions for using echarts4r within Shiny
applications and interactive Rmd documents.
}
