% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{echarts4rBoxOutput}
\alias{echarts4rBoxOutput}
\title{Box Output}
\usage{
echarts4rBoxOutput(id, height = 150)
}
\arguments{
\item{id}{Id of box.}

\item{height}{Height of box, any valid CSS value, numerics are treated as pixels.}
}
\description{
Place box output in Shiny ui.
}
