% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{e_geo3d}
\alias{e_geo3d}
\alias{e_geo3d_lines}
\title{Geo 3D}
\usage{
e_geo3d(e, serie, color, type = "world", ...)

e_geo3d_lines(e, source.lon, source.lat, target.lon, target.lat,
  type = "world", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{color}{Color.}

\item{type}{Map type.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}

\item{name}{name of the serie.}
}
\description{
Create a 3d geo plot.
}
\examples{
choropleth <- data.frame(
  countries = c("France", "Brazil", "China", "Russia", "Canada", "India", "United States",
                "Argentina", "Australia"),
  height = runif(9, 1, 5),
  color = c("#F7FBFF", "#DEEBF7", "#C6DBEF", "#9ECAE1", "#6BAED6", "#4292C6", 
            "#2171B5", "#08519C", "#08306B")
)

choropleth \%>\% 
  e_charts(countries) \%>\% 
  e_geo3d(height, color)
  
flights <- read.csv(
  paste0("https://raw.githubusercontent.com/plotly/datasets/",
         "master/2011_february_aa_flight_paths.csv")
)

choropleth \%>\% 
  e_charts(countries) \%>\% 
  e_map3d(height) \%>\% 
  e_data(flights) \%>\% 
  e_geo3d_lines(
    start_lon, 
    start_lat, 
    end_lon, 
    end_lat,
    name = "flights",
  )

}
