% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_graph}
\alias{e_graph}
\alias{e_graph_nodes}
\alias{e_graph_edges}
\title{Graph}
\usage{
e_graph(e, layout = "force", name = NULL, ...)

e_graph_nodes(e, nodes, names, value, size, category)

e_graph_edges(e, edges, source, target)
}
\arguments{
\item{e}{An \code{echarts4} object as returned by \code{e_charts}.}

\item{name}{Name of graph.}

\item{...}{Any other parameter.}

\item{nodes}{Data.frame of nodes.}

\item{names}{Names of nodes, unique.}

\item{value}{values of nodes.}

\item{size}{Size of nodes.}

\item{category}{Group of nodes (i.e.: group membership).}

\item{edges}{Data.frame of edges.}

\item{source, target}{Column names of source and target.}
}
\description{
Create a graph.
}
\examples{
nodes <- data.frame(
  name = LETTERS,
  value = rnorm(26),
  size = rnorm(26) * 15,
  grp = rep(c("grp1", "grp2"), 13),
  stringsAsFactors = FALSE
)

edges <- data.frame(
  source = sample(LETTERS, 40, replace = TRUE),
  target = sample(LETTERS, 40, replace = TRUE),
  stringsAsFactors = FALSE
)

e_charts() \%>\% 
  e_graph(focusNodeAdjacency = TRUE) \%>\% 
  e_graph_nodes(nodes, name, value, size, grp) \%>\% 
  e_graph_edges(edges, source, target)

}
