% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{highlight_action}
\alias{highlight_action}
\alias{e_highlight}
\alias{e_downplay}
\title{Highlight & Downplay}
\usage{
e_highlight(e, series.index = NULL, series.name = NULL, btn = NULL)

e_downplay(e, series.index = NULL, series.name = NULL, btn = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{series.index}{Index of serie to highlight or \code{list}/\code{vector} of series.}

\item{btn}{A \code{\link{e_button}} id.}
}
\description{
Highlight series
}
\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_charts(Sepal.Length) \%>\% 
  e_line(Sepal.Width) \%>\% 
  e_line(Petal.Length) \%>\% 
  e_highlight(series.name = "setosa") # highlight group

}
