% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{e_show_loading}
\alias{e_show_loading}
\alias{e_hide_loading}
\title{Loading}
\usage{
e_show_loading(e, text = "loading", color = "#c23531",
  text.color = "#000", mask.color = "rgba(255, 255, 255, 0.8)",
  zlevel = 0)

e_hide_loading(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{text}{Text to display.}

\item{color}{Color of spinner.}

\item{text.color}{Color of text.}

\item{mask.color}{Color of mask.}

\item{zlevel}{Z level.}
}
\description{
Show or hide loading.
}
\examples{
mtcars \%>\% 
  e_charts(mpg) \%>\% 
  e_scatter(drat) \%>\% 
  e_show_loading()

}
