% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{map_actions}
\alias{map_actions}
\alias{e_map_select}
\alias{e_map_unselect}
\alias{e_map_toggle_select}
\title{Map Actions}
\usage{
e_map_select(e, ..., btn = NULL)

e_map_unselect(e, ..., btn = NULL)

e_map_toggle_select(e, ..., btn = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any options, see \href{https://ecomfe.github.io/echarts-doc/public/en/api.html#action.map}{official documentation}}

\item{btn}{A \code{\link{e_button}} id.}
}
\description{
Map-related actions.
}
\seealso{
\code{\link{e_map_register}}
}
