% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{pie_action}
\alias{pie_action}
\alias{e_pie_select}
\alias{e_pie_unselect}
\title{Select & Unselect Pie}
\usage{
e_pie_select(e, ..., btn = NULL)

e_pie_unselect(e, ..., btn = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any options, see \href{https://echarts.apache.org/en/api.html#action.pie}{official documentation}}

\item{btn}{A \code{\link{e_button}} id.}
}
\description{
Actions related to \code{\link{e_pie}}.
}
\examples{
mtcars \%>\% 
  head() \%>\% 
  dplyr::mutate(model = row.names(.)) \%>\% 
  e_charts(model) \%>\% 
  e_pie(carb) \%>\% 
  e_pie_select(dataIndex = 0)

}
