% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{renderEcharts4rBox}
\alias{renderEcharts4rBox}
\title{Render box}
\usage{
renderEcharts4rBox(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that produces as \code{\link{echarts4rBox}}.}

\item{env}{The environment in which to evaluate ‘expr’.}

\item{quoted}{Is ‘expr’ a quoted expression (with ‘quote()’)? This is useful if you want to save an expression in a variable.}
}
\description{
Render an echarts4r box.
}
