% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_scatter}
\alias{e_scatter}
\alias{e_effect_scatter}
\title{Scatter}
\usage{
e_scatter(e, serie, size, scale = "* 1", name = NULL,
  coord.system = "cartesian2d", y.index = 0, x.index = 0, ...)

e_effect_scatter(e, serie, size, scale = "* 1", name = NULL,
  coord.system = "cartesian2D", y.index = 0, x.index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{size}{Column name containing size of points.}

\item{scale}{Scale for \code{size}, defaults to \code{* 1} which multiplies the size
by \code{1} (equivalent to no multiplier).}

\item{name}{name of the serie.}

\item{coord.system}{Coordinate system to plot against.}

\item{y.index}{Indexes of x and y axis.}

\item{x.index}{Indexes of x and y axis.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add scatter serie.
}
\examples{
USArrests \%>\% 
  e_charts(Assault) \%>\% 
  e_scatter(Murder, Rape) \%>\% 
  e_scatter(Rape, Murder, y.index = 1) \%>\% 
  e_grid(index = c(0, 1)) \%>\% 
  e_tooltip()

quakes \%>\% 
  e_charts(long) \%>\% 
  e_geo(
    roam = TRUE,
    boundingCoords = list(
      c(185, - 10),
      c(165, -40)
    )
  ) \%>\% 
  e_effect_scatter(lat, mag, coord.system = "geo") \%>\% 
  e_visual_map(min = 4, max = 6.5)
  
# Calendar
year1 <- seq.Date(as.Date("2018-01-01"), as.Date("2018-12-31"), by = "day")
year2 <- seq.Date(as.Date("2017-01-01"), as.Date("2017-12-31"), by = "day")
values <- rnorm(length(dates), 20, 6)

year <- data.frame(year1 = year1, year2 = year2, values = values)

year \%>\% 
  e_charts(year1) \%>\% 
  e_calendar(range = "2017") \%>\% 
  e_scatter(values, coord.system = "calendar", calendarIndex = 0) \%>\% 
  e_data(year, year2) \%>\% 
  e_calendar(range = "2018", top = 300) \%>\% 
  e_scatter(values, coord.system = "calendar", calendarIndex = 1) \%>\% 
  e_visual_map(max = 30)

}
