% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_lm}
\alias{e_lm}
\alias{e_glm}
\alias{e_loess}
\title{Smooth}
\usage{
e_lm(e, formula, name = NULL, legend = TRUE, symbol = "none", ...)

e_glm(e, formula, name = NULL, legend = TRUE, symbol = "none", ...)

e_loess(e, formula, name = NULL, legend = TRUE, symbol = "none", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{formula}{formula to pass to \code{\link{lm}}.}

\item{name, }{series name.}

\item{legend}{Whether to plot legend.}

\item{symbol}{Symbol to use in \code{\link{e_line}}.}

\item{...}{Additional arguments to pass to \code{\link{lm}}, \code{\link{glm}}, 
and \code{\link{loess}}.}
}
\description{
Aids the eye in seeing patterns in the presence of overplotting.
}
\examples{
mtcars \%>\% 
  e_charts(mpg) \%>\% 
  e_scatter(qsec) \%>\% 
  e_lm(qsec ~ mpg, name = "y = ax + b")
  
mtcars \%>\% 
  e_charts(disp) \%>\% 
  e_scatter(mpg) \%>\% 
  e_loess(mpg ~ disp)
  
CO2 \%>\% 
  e_charts(conc) \%>\% 
  e_scatter(uptake) \%>\% 
  e_glm(uptake ~ conc, name = "GLM")

}
