% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline-serie}
\alias{timeline-serie}
\alias{e_timeline_serie}
\alias{e_timeline_serie_}
\title{Timeline serie}
\usage{
e_timeline_serie(e, y, z, type = "line", name = NULL,
  serie_name = NULL, legend = FALSE, base_opts = NULL,
  coord_system = "cartesian2d")

e_timeline_serie_(e, y, z = NULL, type = "line", name = NULL,
  serie_name = NULL, legend = FALSE, base_opts = NULL,
  coord_system = "cartesian2d")
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_timeline}}.}

\item{y, z}{Any column name.}

\item{type}{Chart type.}

\item{name}{name of serie.}

\item{serie_name}{Name of serie, used for legend.}

\item{legend}{Set to \code{TRUE} to show \code{serie_name} in legend.}

\item{base_opts}{A \code{list} of base serie options.}
}
\description{
Add timeline series.
}
\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_timeline(Sepal.Length, autoPlay = TRUE) \%>\% 
  e_timeline_serie(
    Sepal.Width, 
    serie_name = "Width", 
    legend = TRUE
  ) \%>\% 
  e_timeline_serie(
    Petal.Length, 
    serie_name = "Length", 
    legend = TRUE
  ) 

}
