% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_toolbox_feature}
\alias{e_toolbox_feature}
\alias{e_toolbox}
\title{Toolbox}
\usage{
e_toolbox_feature(e, feature, ...)

e_toolbox(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{feature}{Feature to add, defaults to all.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Add toolbox interface.
}
\examples{
USArrests \%>\% 
  e_charts(UrbanPop) \%>\% 
  e_line(Assault) \%>\% 
  e_area(Murder, y.index = 1, x.index = 1) \%>\% 
  e_datazoom(x.index = 0) 

}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#toolbox}{Additional arguments}
}
