% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{tooltip_action}
\alias{tooltip_action}
\alias{e_showtip}
\alias{e_hidetip}
\title{Show & Hide Tooltip}
\usage{
e_showtip(e, ..., btn = NULL)

e_hidetip(e, ..., btn = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{...}{Any options, see \href{https://echarts.apache.org/en/api.html#action.tooltip}{official documentation}}

\item{btn}{A \code{\link{e_button}} id.}
}
\description{
Show or hide tooltip.
}
\note{
The tooltip must be initialised with \code{\link{e_tooltip}} for this to work.
}
\examples{
cars \%>\%
  e_charts(dist) \%>\%
  e_scatter(speed) \%>\%
  e_tooltip() \%>\%
  e_hidetip(btn = "btn") \%>\%
  e_button("btn", "Hide tooltip")
}
