% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3d.R
\name{e_axis_3d}
\alias{e_axis_3d}
\alias{e_x_axis_3d}
\alias{e_y_axis_3d}
\alias{e_z_axis_3d}
\title{Axis 3D}
\usage{
e_axis_3d(e, axis = c("x", "y", "z"), index = 0, ...)

e_x_axis_3d(e, index = 0, ...)

e_y_axis_3d(e, index = 0, ...)

e_z_axis_3d(e, index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{axis}{Axis to customise.}

\item{index}{Index of axis to customise.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise 3D axis.
}
