% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{connections}
\alias{connections}
\alias{e_connect}
\alias{e_arrange}
\title{Connect charts}
\usage{
e_connect(e, ids)

e_arrange(..., rows = NULL, cols = NULL, title = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{ids}{Scalar, vector or list of ids of chart to connect with.}

\item{...}{Any \code{echarts} objects.}

\item{rows, cols}{Number of rows and columns.}

\item{title}{Title of charts.}
}
\value{
In an interactive session returns a \link[htmltools]{browsable}, in \code{rmarkdown} returns a 
container (\link[htmltools]{div}).
}
\description{
Connect charts together.
}
\examples{
# linked datazoom
e1 <- cars \%>\% 
  e_charts(
    speed,
    height = 200,
    elementId = "chart1"
  ) \%>\% 
  e_scatter(dist) \%>\% 
  e_datazoom(show = FALSE)
  
e2 <- cars \%>\% 
  e_charts(
    dist,
    height = 200
  ) \%>\% 
  e_scatter(speed) \%>\% 
  e_datazoom() \%>\% 
  e_connect("chart1") # connect

e_arrange(e1, e2, title = "Linked datazoom")

}
