% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{e_annotations}
\alias{e_annotations}
\title{Add annotations to a chart}
\usage{
e_annotations(e, annotations, grid = NULL, series = NULL)
}
\arguments{
\item{e}{An echarts4r object}

\item{annotations}{list of annotations to plot}
}
\description{
Apache ECharts does not include a native annotation system. This is a custom
function that creates annotations using ECharts’ low-level graphic components
and SVG for the line. This currently only works in a 'cartesian2d' coordinate
system.

Each annotation must be in a list with an x, y, and text. Styling can be
added - see @details.

In Shiny, to output an annotation position after dragging the box, use
\code{input$id_dragged_annotation} or see \link{echarts4r-shiny}. This
captures the annotation parameters - not any of the styles.
}
\details{
annotations can take the following styles to change the defaults. To
  remove any element use "none".

- \strong{group}: Controls the box and text elements. color was added as an option. This color colors the text, box border, line and arrow - unless specified in that particular style argument.

- \strong{rectStyle} Styles the annotation box.

- \strong{textStyle}, Styles the annotation text. textAlign and padding was added.

- \strong{lineStyle} Styles the line that connects the annotation box to the arrow using \href{https://www.w3schools.com/graphics/svg_stroking.asp}{SVG stroke attributes}

- \strong{arrowStyle}: Styles the arrow. size was added.
}
\examples{
mtcars |>
e_charts(mpg) |>
  e_scatter(wt) |>
  e_annotations(
    annotations = list(
      list(
        id = 0,
        x = 15,
        y = 3,
        text = 'An annotation\nwith styles',
        offsetX = 0,
        offsetY = -50,
        # Using the styles
        group = list(draggable = FALSE, color = "red"),
        textStyle = list(
          "font-size" = 14,
          "font-weight" = 'bold',
          color = "green"
        ),
        rectStyle = list(
          lineDash = 'dashed',
          lineWidth = 2,
          shape = list(width = 105, height = 35, r = 0)
        ),
        lineStyle = list(`stroke-width` = 2),
        arrowStyle = list(size = 8)
      ),
      # No styles, will just show text
      list(
        id = 0,
        x = 25,
        y = 1,
        text = 'No styles',
        offsetX = 0,
        offsetY = -40,
        lineStyle = "none",
        rectStyle = "none",
        arrowStyle = "none",
        textStyle = list(color = "red")
      ),
      # Style text using rich text
      list(
        id = 2,
        x = 25,
        y = 4.5,
        text = "{bold|Text using}\n{red|rich text!}",
        offsetX = 0,
        offsetY = -40,
        textStyle = list(
          rich = list(
            bold = list(fontWeight = 'bold'),
            red = list(fill = 'red')
          )
        )
      )
    ))

}
\seealso{
- \href{https://developer.mozilla.org/en-US/docs/Web/SVG/Reference/Attribute}{Additional arguments for SVG attributes}

- \href{https://echarts.apache.org/en/option.html#graphic.elements-text.style}{Additional arguments for textStyle}

- \href{https://www.w3schools.com/graphics/svg_stroking.asp}{Additional arguments for rectStyle}

- \href{https://www.w3schools.com/graphics/svg_stroking.asp}{Additional
arguments for lineStyle}

- \href{https://echarts.apache.org/en/option.html#graphic.elements-polygon.style}{Additional
arguments for arrowStyle}
}
