% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{e_annotations}
\alias{e_annotations}
\title{Add annotations to a chart}
\usage{
e_annotations(e, annotations)
}
\arguments{
\item{e}{e}

\item{annotations}{list of annotation to plot}
}
\description{
Each annotation requires an x, y, and text.
Only works with cartesian2d.
 In Shiny, to output an annotation position, use `input$annotation_positions`
 For rectStyle if x and y are ignored then line is connected on the horizontal bottom of the box.
For textStyle, rich text does work and will be applied to annotation text. textAlign and padding was also added.
For arrowStyle, size was added.
Make any "none" to remove element.
If color is provided in group, the text, box border, line and arrow will use that color.
}
\details{
- [ECharts graphics component](https://echarts.apache.org/en/option.html#graphic)
- [ECharts-GL globe](https://echarts.apache.org/en/option-gl.html#globe)

- for textStyle list of [text.style](https://echarts.apache.org/en/option.html#graphic.elements-text.style)

- list of [group](https://echarts.apache.org/en/option.html#graphic.elements-group)

- rectStyle list of [rect.style](https://echarts.apache.org/en/option.html#graphic.elements-rect.style)

- lineStyle SVG list for [lines](https://www.w3schools.com/graphics/svg_stroking.asp)

- arrowStyle list of [polygon.style](https://echarts.apache.org/en/option.html#graphic.elements-polygon.style)
}
\examples{
mtcars |>
e_charts(mpg) |>
  e_scatter(wt) |>
  e_annotations(
    annotations = list(
      list(
        id = 0,
        x = 15,
        y = 3,
        text = 'An annotation',
        offsetX = 0,
        offsetY = -50,
        group = list(draggable = FALSE, color = "red"),
        textStyle = list(
          fontSize = 14,
          fontWeight = 'bold',
          color = "green"
        ),
        rectStyle = list(
          lineDash = 'solid',
          lineWidth = 2,
          shape = list(width = 105, height = 35, r = 0)
        ),
        lineStyle = list(`stroke-width` = 2),
        arrowStyle = list(size = 8)
      ),
      # No styles, will just show text
      list(
        id = 0,
        x = 25,
        y = 1,
        text = 'No Styles',
        offsetX = 0, # box position from x,y
        offsetY = -40,
        color = "red",
        lineStyle = "none",
        rectStyle = "none",
        arrowStyle = "none"
      ),
      list(
        id = 2,
        x = 25,
        y = 4.5,
        text = "{bold|Text using}\n{red|rich text!}",
        offsetX = 0, # box position
        offsetY = -40,
        textStyle = list(
          rich = list(
            bold = list(fontWeight = 'bold'),
            red = list(fill = 'red')
          )
        ),
        rectStyle = list(
          lineDash = 'solid'
        )
      )
    ))
}
