% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_area}
\alias{e_area}
\title{Area}
\usage{
e_area(e, serie, name = NULL, y.index = 0, x.index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{y.index}{Indexes of x and y axis.}

\item{x.index}{Indexes of x and y axis.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add area serie.
}
\examples{
USArrests \%>\% 
  e_charts(Assault) \%>\% 
  e_area(Murder, stack = "grp") \%>\% # stacking
  e_area(UrbanPop, stack = "grp") # stacking

}
