% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_axis_pointer}
\alias{e_axis_pointer}
\title{Axis pointer}
\usage{
e_axis_pointer(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Customise axis pointer.
}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#axisPointer}{Additional arguments}
}
