% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_bar}
\alias{e_bar}
\title{Bar & Line chart}
\usage{
e_bar(e, serie, name = NULL, y.index = 0, x.index = 0, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add bar serie.
}
\examples{
USArrests \%>\% 
  dplyr::mutate(
    State = row.names(.),
    Rape = -Rape
  ) \%>\% 
  e_charts(State) \%>\% 
  e_bar(Murder) \%>\% 
  e_bar(Rape, name = "Sick basterd", x.index = 1) # second y axis

}
