% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_bar_3d}
\alias{e_bar_3d}
\title{Bars 3D}
\usage{
e_bar_3d(e, y, z, coord.system = "cartesian3D", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{y, z}{Coordinates.}

\item{coord.system}{Coordinate system to use, one of \code{cartesian3D}, \code{geo3D}, \code{globe}.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add 3D bars
}
\examples{
url <- paste0("https://ecomfe.github.io/echarts-examples/",
              "public/data-gl/asset/data/population.json")
data <- jsonlite::fromJSON(url)
data <- as.data.frame(data)
names(data) <- c("lon", "lat", "value")

data \%>\% 
  e_charts(lon) \%>\% 
  e_globe(
    environment = e_stars_texture(),
    base.texture = e_globe_texture()
  ) \%>\% 
  e_bar_3d(lat, value, coord.system = "globe") \%>\% 
  e_visual_map()
  
v <- LETTERS[1:10]
matrix <- data.frame(
  x = sample(v, 300, replace = TRUE), 
  y = sample(v, 300, replace = TRUE), 
  z = rnorm(300, 10, 1),
  stringsAsFactors = FALSE
) \%>\% 
  dplyr::group_by(x, y) \%>\% 
  dplyr::summarise(z = sum(z)) \%>\% 
  dplyr::ungroup() 
  
matrix \%>\% 
  e_charts(x) \%>\% 
  e_bar_3d(y, z)

}
