% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{e_bar_3d}
\alias{e_bar_3d}
\alias{e_bar_3d_}
\title{Bar 3D}
\usage{
e_bar_3d(
  e,
  y,
  z,
  bind,
  coord_system = "cartesian3D",
  name = NULL,
  rm_x = TRUE,
  rm_y = TRUE,
  ...
)

e_bar_3d_(
  e,
  y,
  z,
  bind = NULL,
  coord_system = "cartesian3D",
  name = NULL,
  rm_x = TRUE,
  rm_y = TRUE,
  ...
)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{y, z}{Coordinates.}

\item{bind}{Binding.}

\item{coord_system}{Coordinate system to use, one of \code{cartesian3D}, \code{geo3D}, \code{globe}.}

\item{name}{name of the serie.}

\item{rm_x, rm_y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Add 3D bars
}
\examples{
\dontrun{
# volcano
volcano \%>\% 
  as.table() \%>\% 
  as.data.frame() \%>\% 
  dplyr::mutate(
    Var1 = as.integer(Var1),
    Var2 = as.integer(Var2)
  ) \%>\% 
  e_charts(Var1) \%>\% 
  e_bar_3d(Var2, Freq) \%>\% 
  e_visual_map(Freq)

url <- paste0("https://ecomfe.github.io/echarts-examples/",
              "public/data-gl/asset/data/population.json")
data <- jsonlite::fromJSON(url)
data <- as.data.frame(data)
names(data) <- c("lon", "lat", "value")

# globe
data \%>\% 
  e_charts(lon) \%>\% 
  e_globe() \%>\% 
  e_bar_3d(lat, value, coord_system = "globe") \%>\% 
  e_visual_map()

# get3d
data \%>\% 
  e_charts(lon) \%>\% 
  e_geo_3d() \%>\% 
  e_bar_3d(lat, value, coord_system = "geo3D") \%>\% 
  e_visual_map()

# stacked
v <- LETTERS[1:10]
matrix <- data.frame(
  x = sample(v, 300, replace = TRUE), 
  y = sample(v, 300, replace = TRUE), 
  z1 = rnorm(300, 10, 1),
  z2 = rnorm(300, 10, 1),
  stringsAsFactors = FALSE
) \%>\% 
  dplyr::group_by(x, y) \%>\% 
  dplyr::summarise(
    z1 = sum(z1),
    z2 = sum(z2)
  ) \%>\% 
  dplyr::ungroup() 
  
trans <- list(opacity = 0.4) # transparency
emphasis <- list(itemStyle = list(color = "#313695"))
  
matrix \%>\% 
  e_charts(x) \%>\% 
  e_bar_3d(y, z1, stack = "stack", name = "Serie 1", itemStyle = trans, emphasis = emphasis) \%>\%
  e_bar_3d(y, z2, stack = "stack", name = "Serie 2", itemStyle = trans, emphasis = emphasis) \%>\% 
  e_legend()
  
# timeline
matrix \%>\% 
  group_by(x) \%>\% 
  e_charts(y, timeline = TRUE) \%>\% 
  e_bar_3d(z1, z2) \%>\% 
  e_visual_map(z2)
}

}
\seealso{
\href{http://echarts.baidu.com/option-gl.html#series-bar3D}{Additional arguments}
}
