% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_calendar}
\alias{e_calendar}
\title{Calendar}
\usage{
e_calendar(e, serie, range, ...)
}
\description{
Calendar
}
\examples{
dates <- seq.Date(
  as.Date("2017-01-01"), 
  as.Date("2017-12-31"), 
  by = "day"
)

calendar <- data.frame(
  day = dates,
  value = rnorm(length(dates))
)

calendar \%>\% 
  e_chart(day) \%>\% 
  e_calendar(value, range = "2017")

}
