% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_calendar}
\alias{e_calendar}
\title{Calendar}
\usage{
e_calendar(e, serie, range, name = NULL, ...)
}
\description{
Calendar
}
\examples{
dates <- seq.Date(
  as.Date("2017-01-01"), 
  as.Date("2017-12-30"), 
  by = "day"
)

val <- rnorm(length(dates), 20, 3)

calendar <- data.frame(
  day = dates,
  value = val
)

calendar \%>\% 
  e_charts(day) \%>\% 
  e_calendar(value, range = "2017") \%>\% 
  e_visual_map(
    min = min(val),
    max = max(val)
  ) \%>\% 
  e_tooltip()

}
