% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_calendar}
\alias{e_calendar}
\title{Calendar}
\usage{
e_calendar(e, serie, range, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{range}{Range of calendar.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Heatmap as calendar.
}
\examples{
dates <- seq.Date(
  as.Date("2017-01-01"), 
  as.Date("2017-12-30"), 
  by = "day"
)

val <- rnorm(length(dates), 20, 3)

calendar <- data.frame(
  day = dates,
  value = val
)

calendar \%>\% 
  e_charts(day) \%>\% 
  e_calendar(value, range = "2017") \%>\% 
  e_visual_map(
    min = min(val),
    max = max(val)
  ) \%>\% 
  e_tooltip()

}
