% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{e_calendar}
\alias{e_calendar}
\title{Calendar}
\usage{
e_calendar(e, range, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{range}{Range of calendar format, string or vector.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Calendar
}
\examples{
# year
mtcars \%>\% 
  e_charts() \%>\% 
  e_calendar(range = "2017")
  
# month
mtcars \%>\% 
  e_charts() \%>\% 
  e_calendar(range = "2018-01")
  
# range
mtcars \%>\% 
  e_charts() \%>\% 
  e_calendar(range = c("2018-01", "2018-07"))

}
\seealso{
\href{https://echarts.apache.org/en/option.html#calendar}{Additional arguments}
}
