% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{e_choropleth}
\alias{e_choropleth}
\title{Choropleth}
\usage{
e_choropleth(e, serie, region, type = "world", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Values to plot.}

\item{region}{Names of regions/countries or states.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw a choropleth map.
}
\examples{
choropleth <- data.frame(
  countries = c("France", "Brazil", "China", "Russia", "Canada", "India"),
  values = round(runif(6, 10, 25))
)

choropleth \%>\% 
  e_charts() \%>\% 
  e_choropleth(values, countries)

}
