% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{e_color}
\alias{e_color}
\title{Color}
\usage{
e_color(e, color = NULL, background = NULL)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{color}{Vector of colors.}

\item{background}{Background color.}
}
\description{
Customise chart and background colors.
}
\examples{
mtcars \%>\% 
  e_charts(drat) \%>\% 
  e_line(mpg) \%>\% 
  e_area(qsec) \%>\% 
  e_color(
    c("red", "blue"),
    "#d3d3d3"
  )
  
}
\seealso{
\code{\link{e_theme}}, 
\href{https://echarts.apache.org/en/option.html#color}{Official color documentation}, 
\href{https://echarts.apache.org/en/option.html#backgroundColor}{Official background documentation}
}
