% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_connect}
\alias{e_connect}
\title{Set Group}
\usage{
e_connect(e, ids)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{ids}{Scalar, vector or list of ids of chart to connect with.}
}
\description{
Set group to connect two charts together.
}
\examples{
library(htmltools)
  
# linked datazoom
e1 <- cars \%>\% 
  e_charts(
    speed,
    elementId = "chart1"
  ) \%>\% 
  e_scatter(dist) \%>\% 
  e_datazoom(show = FALSE)
  
e2 <- cars \%>\% 
  e_charts(
    dist
  ) \%>\% 
  e_scatter(speed) \%>\% 
  e_datazoom() \%>\% 
  e_connect("chart1") # connect

browsable(
  div(
    div(e1),
    div(e2)
  )
)

}
