% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_dims}
\alias{e_dims}
\title{Dimensions}
\usage{
e_dims(e, height = "auto", width = "auto")
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{height, width}{Dimensions in pixels, percentage or string.}
}
\description{
Sets the dimensions of the chart _internally._
This will only affect the dimensions of the
chart within its parent container.
Use the `height` and `width` arguments of 
[e_charts] if you want to change the dimensions
of said parent (recommended).
}
