% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{e_execute}
\alias{e_execute}
\alias{e_execute_p}
\title{Send}
\usage{
e_execute(proxy)

e_execute_p(proxy)
}
\arguments{
\item{proxy}{An echarts4r proxy as returned by \code{\link{echarts4rProxy}}.}
}
\description{
Send new series to chart.
}
