% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{e_facet}
\alias{e_facet}
\title{Facet}
\usage{
e_facet(e, rows = 1, cols = 1, legend_pos = "top", legend_space = 10)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{rows, cols}{Number of rows and columns.}

\item{legend_pos}{Position of the legend}

\item{legend_space}{Space between legend and plot area.}
}
\description{
Create facets for multiple plots.
}
\details{
Each serie, i.e.: \code{\link{e_bar}} will be plotted against a facet.
}
