% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_gauge}
\alias{e_gauge}
\alias{e_gauge_}
\title{Gauge}
\usage{
e_gauge(e, value, name, rm_x = TRUE, rm_y = TRUE, ...)

e_gauge_(e, value, name, rm_x = TRUE, rm_y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{value}{Value to gauge.}

\item{name}{Text on gauge.}

\item{rm_x, rm_y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Plot a gauge.
}
\examples{
e_charts() \%>\% 
  e_gauge(57, "PERCENT")

}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#series-gauge}{Additional arguments}
}
