% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{e_get_data}
\alias{e_get_data}
\title{Get data}
\usage{
e_get_data(e)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or 
a proxy as returned by \code{\link{echarts4rProxy}}.}
}
\value{
A list of data.frames, one for each group.
}
\description{
Get data passed to \code{\link{e_charts}}.
}
\examples{
echart <- cars \%>\% 
  e_charts(speed) \%>\% 
  e_scatter(dist) \%>\% 
  e_lm(dist ~ speed) 
  
echart

e_get_data(echart)[[1]]

}
