% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe.R
\name{e_globe}
\alias{e_globe}
\title{Globe}
\usage{
e_globe(e, environment = NULL, height.texture = NULL, base.texture = NULL,
  ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add globe.
}
\examples{
\dontrun{
url <- paste0("https://ecomfe.github.io/echarts-examples/",
              "public/data-gl/asset/data/population.json")
data <- jsonlite::fromJSON(url)
data <- as.data.frame(data)
names(data) <- c("lon", "lat", "value")

data \%>\% 
  e_charts() \%>\% 
  e_globe(
    environment = e_stars_texture(),
    base.texture = e_globe_texture()
  ) \%>\% 
  e_bar_3d(lon, lat, value, "globe") \%>\% 
  e_visual_map()
}

}
