% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_heatmap}
\alias{e_heatmap}
\title{Heatmap}
\usage{
e_heatmap(e, x, y, z, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4} object as returned by \code{e_charts}.}

\item{x, y, z}{Coordinates and values.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw heatmap by coordinates.
}
\examples{
v <- LETTERS[1:10]
matrix <- data.frame(
  x = sample(v, 300, replace = TRUE), 
  y = sample(v, 300, replace = TRUE), 
  z = rnorm(300, 10, 1)
) \%>\% 
  dplyr::group_by(x, y) \%>\% 
  dplyr::summarise(z = sum(z)) \%>\% 
  dplyr::ungroup()

matrix \%>\% 
  e_charts() \%>\% 
  e_heatmap(x, y, z) \%>\% 
  e_visual_map()

}
