% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_line}
\alias{e_line}
\title{Line}
\usage{
e_line(e, serie, name = NULL, y.index = 0, x.index = 0,
  coord.system = "cartesian2d", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{y.index}{Indexes of x and y axis.}

\item{x.index}{Indexes of x and y axis.}

\item{coord.system}{Coordinate system to plot against.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add line serie.
}
\examples{
USArrests \%>\% 
  e_charts(Assault) \%>\% 
  e_line(Murder) \%>\% 
  e_line(UrbanPop, y.index = 1) # second y axis

}
