% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_lines}
\alias{e_lines}
\title{Lines}
\usage{
e_lines(e, source.lon, source.lat, target.lon, target.lat,
  coord.system = "geo", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{source.lon, source.lat, target.lon, target.lat}{coordinates.}

\item{coord.system}{Coordinate system to use, one of \code{geo}, or \code{cartesian2D}.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Add lines.
}
\examples{
flights <- read.csv(
  paste0("https://raw.githubusercontent.com/plotly/datasets/",
         "master/2011_february_aa_flight_paths.csv")
)

flights \%>\% 
  e_charts() \%>\% 
  e_geo() \%>\% 
  e_lines(
    start_lon, 
    start_lat, 
    end_lon, 
    end_lat,
    name = "flights",
    lineStyle = list(normal = list(curveness = 0.3))
   )

}
