% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_lines_gl}
\alias{e_lines_gl}
\title{Lines WebGL}
\usage{
e_lines_gl(e, data, coord_system = "geo", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{data}{A list.}

\item{coord_system}{Coordinate system to plot against.}

\item{...}{Any other options (this series type is mostly undocumented).}
}
\description{
Draw WebGL lines.
}
