% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{e_matrix}
\alias{e_matrix}
\title{Generate Matrix}
\usage{
e_matrix(e, xAxis, yAxis, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{xAxis, yAxis}{provide column name of dataframe to generate X-axis and Y-axis header cells}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
helper function for generating the x and y axes for a matrix grid.
}
\examples{

df <- data.frame("Class" = rep(c("Class1", "Class2", "Class3"),each = 3),
"Grade" = c("Grade1","Grade2", "Grade3"),
"Males" = sample(1:10, 9),
"Females" = sample(1:10,9))

df |> e_charts() |> e_matrix(xAxis = "Class", yAxis = "Grade")

}
\seealso{
\href{https://echarts.apache.org/en/option.html#matrix}{Additional arguments}
}
