% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{e_matrix_pie}
\alias{e_matrix_pie}
\title{Generate pie chart for matrix}
\usage{
e_matrix_pie(e, x, y, legend = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}} or
a proxy as returned by \code{\link{echarts4rProxy}}.}

\item{x, y}{x and y column names for data to be used in the pie chart}

\item{legend}{Whether to add a legend, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Draw pie or donut chart in matrix coordinate system
}
\seealso{
\href{https://echarts.apache.org/en/option.html#series-pie}{Additional arguments}
}
