% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_pie}
\alias{e_pie}
\title{Pie}
\usage{
e_pie(e, serie, name = NULL, rm.x = TRUE, rm.y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{rm.x, rm.y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Draw pie and donut charts.
}
\examples{
mtcars \%>\% 
  head() \%>\% 
  dplyr::mutate(model = row.names(.)) \%>\% 
  e_charts(model) \%>\% 
  e_pie(carb)
  
}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#series-pie}
}
