% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_radar}
\alias{e_radar}
\title{Radar}
\usage{
e_radar(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise radar.
}
