% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{e_radius_axis}
\alias{e_radius_axis}
\title{Radius axis}
\usage{
e_radius_axis(e, show = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{show}{Whether to display the axis.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Customise radius axis.
}
\examples{
df <- data.frame(x = 1:10, y = seq(1, 20, by = 2))

df \%>\% 
  e_charts(x) \%>\% 
  e_polar() \%>\% 
  e_angle_axis() \%>\% 
  e_radius_axis(FALSE) \%>\% 
  e_line(y, coord.system = "polar", smooth = TRUE) 

}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#radiusAxis}
}
