% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_river}
\alias{e_river}
\title{River}
\usage{
e_river(e, serie, name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Column name of serie to plot.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Build a theme river.
}
\examples{
dates <- seq.Date(Sys.Date() - 30, Sys.Date(), by = "day")

df <- data.frame(
  dates = dates,
  apples = runif(length(dates)),
  bananas = runif(length(dates)),
  pears = runif(length(dates))
)

df \%>\% 
  e_charts(dates) \%>\% 
  e_river(apples) \%>\% 
  e_river(bananas) \%>\% 
  e_river(pears)

}
