% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_sankey}
\alias{e_sankey}
\title{Sankey}
\usage{
e_sankey(e, source, target, value, layout = "none", ...)
}
\description{
Sankey
}
\examples{
sankey <- data.frame(
  source = sample(LETTERS[1:10], 20, replace = TRUE),
  target = sample(LETTERS[1:10], 20, replace = TRUE),
  value = rnorm(10)
)

sankey \%>\% 
  e_charts() \%>\% 
  e_sankey(source, target, value)

}
