% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_sankey}
\alias{e_sankey}
\title{Sankey}
\usage{
e_sankey(e, source, target, value, layout = "none", ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{source, target}{Source and target columns.}

\item{value}{Value change from \code{source} to \code{target}.}

\item{layout}{Layout of sankey.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw a sankey diagram.
}
\examples{
sankey <- data.frame(
  source = c("a", "b", "c", "d", "c"),
  target = c("b", "c", "d", "e", "e"),
  value = ceiling(rnorm(5, 10, 1)),
  stringsAsFactors = FALSE
)

sankey \%>\% 
  e_charts() \%>\% 
  e_sankey(source, target, value) 

}
