% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{e_scatter_gl}
\alias{e_scatter_gl}
\alias{e_scatter_gl_}
\title{Scatter GL}
\usage{
e_scatter_gl(e, y, z, name = NULL, coord_system = "geo", rm_x = TRUE,
  rm_y = TRUE, ...)

e_scatter_gl_(e, y, z, name = NULL, coord_system = "geo",
  rm_x = TRUE, rm_y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{y, z}{Column names containing y and z data.}

\item{name}{name of the serie.}

\item{coord_system}{Coordinate system to plot against.}

\item{rm_x, rm_y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Draw scatter GL.
}
\examples{
quakes \%>\% 
  e_charts(long) \%>\% 
  e_geo(
    roam = TRUE,
    boundingCoords = list(
      c(185, - 10),
      c(165, -40)
     )
  ) \%>\% 
  e_scatter_gl(lat, depth)

}
\seealso{
\href{http://echarts.baidu.com/option-gl.html#series-scatterGL}{Additional arguments}
}
