% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_sunburst}
\alias{e_sunburst}
\title{Sunburst}
\usage{
e_sunburst(e, value, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{value}{Name of column containing values.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Build a sunburst.
}
