% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_sunburst}
\alias{e_sunburst}
\title{Sunburst}
\usage{
e_sunburst(e, parent, child, value, rm.x = TRUE, rm.y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{parent, child}{Edges.}

\item{value}{Name of column containing values.}

\item{rm.x, rm.y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Build a sunburst.
}
\examples{
\dontrun{
df <- data.frame(
  parent = c("earth", "earth", "earth", "mars", "mars"), 
  child = c("forest", "ocean", "iceberg", "elon", "curiosity"),
  value = ceiling(rnorm(5, 10, 2))
)

df \%>\% 
  e_charts() \%>\% 
  e_sunburst(parent, child, value) \%>\% 
  e_theme("westeros")
}

}
