% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{e_timeline_opts}
\alias{e_timeline_opts}
\title{Timeline}
\usage{
e_timeline_opts(e, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Set timeline options
}
\examples{
iris \%>\% 
  group_by(Species) \%>\% 
  e_charts(Sepal.Length, timeline = TRUE) \%>\% 
  e_line(Sepal.Width) \%>\% 
  e_timeline_opts(
    autoPlay = TRUE,
    rewind = TRUE
  )
  
}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#timeline}{official documentation}
}
