% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{e_tooltip}
\alias{e_tooltip}
\title{Tooltip}
\usage{
e_tooltip(e, trigger = c("item", "axis"), ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{trigger}{What triggers the tooltip, one of \code{item} or \code{item}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Customise tooltip
}
\examples{
USArrests \%>\% 
  e_charts(Assault) \%>\% 
  e_bar(Murder) \%>\% 
  e_tooltip()

}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#tooltip}{Additional arguments}
}
