% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_tree}
\alias{e_tree}
\title{Tree}
\usage{
e_tree(e, parent, child, rm.x = TRUE, rm.y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{parent, child}{Edges.}

\item{rm.x, rm.y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Build a tree.
}
\examples{
df <- data.frame(
  parent = c("earth","earth","forest","forest","ocean","ocean","ocean","ocean"), 
  child = c("ocean","forest","tree","sasquatch","fish","seaweed","mantis shrimp","sea monster")
)

df \%>\% 
  e_charts() \%>\% 
  e_tree(parent, child)

}
\seealso{
\href{Additional arguments}{https://ecomfe.github.io/echarts-doc/public/en/option.html#series-tree}
}
