% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{e_tree}
\alias{e_tree}
\alias{e_tree_}
\title{Tree}
\usage{
e_tree(e, parent, child, rm_x = TRUE, rm_y = TRUE, ...)

e_tree_(e, parent, child, rm_x = TRUE, rm_y = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{parent, child}{Edges.}

\item{rm_x, rm_y}{Whether to remove x and y axis, defaults to \code{TRUE}.}

\item{...}{Any other option to pass, check See Also section.}
}
\description{
Build a tree.
}
\examples{
df <- data.frame(
  parent = c("earth","earth","forest","forest","ocean","ocean","ocean","ocean"), 
  child = c("ocean","forest","tree","sasquatch","fish","seaweed","mantis shrimp","sea monster")
)

df \%>\% 
  e_charts() \%>\% 
  e_tree(parent, child)

}
\seealso{
\href{https://ecomfe.github.io/echarts-doc/public/en/option.html#series-tree}{Additional arguments}
}
