% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{e_treemap}
\alias{e_treemap}
\title{Tree}
\usage{
e_treemap(e, parent, child, ...)
}
\arguments{
\item{e}{An \code{echarts4} object as returned by \code{e_charts}.}

\item{parent, child}{Edges.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Build a treemap.
}
\examples{
df <- data.frame(
  parent = c("earth","earth","forest","forest","ocean","ocean","ocean","ocean"), 
  child = c("ocean","forest","tree","sasquatch","fish","seaweed","mantis shrimp","sea monster")
)

df \%>\% 
  e_charts() \%>\% 
  e_treemap(parent, child)

}
