% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts4r.R
\name{e_charts}
\alias{e_charts}
\alias{e_charts_}
\alias{e_chart}
\alias{e_data}
\title{Initialise}
\usage{
e_charts(data, x, width = NULL, height = NULL, elementId = NULL,
  dispose = TRUE, renderer = "canvas", ...)

e_charts_(data, x = NULL, width = NULL, height = NULL,
  elementId = NULL, dispose = TRUE, renderer = "canvas", ...)

e_chart(data, x, width = NULL, height = NULL, elementId = NULL,
  dispose = TRUE, renderer = "canvas", ...)

e_data(e, data, x)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x}{Column name containing x axis.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}

\item{dispose}{Set to \code{TRUE} to force redraw of chart, set to \code{FALSE} to update.}

\item{renderer}{Renderer, takes \code{canvas} (default) or \code{svg}.}

\item{...}{Any other argument.}

\item{e}{An object of class \code{echarts4r} as returned by \code{e_charts}.}
}
\description{
Initialise a chart.
}
\examples{
mtcars \%>\% 
  e_charts_("qsec") \%>\%
  e_line(mpg)

}
