% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{e_choropleth}
\alias{e_choropleth}
\alias{e_map_3d}
\title{Choropleth}
\usage{
e_choropleth(e, serie, map = "world", name = NULL, ...)

e_map_3d(e, serie, map = "world", name = NULL, ...)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{serie}{Values to plot.}

\item{map}{Map type.}

\item{name}{name of the serie.}

\item{...}{Any other option to pass to \code{bar} or \code{line} char types.}
}
\description{
Draw maps.
}
\examples{
\dontrun{
choropleth <- data.frame(
  countries = c("France", "Brazil", "China", "Russia", "Canada", "India", "United States",
                "Argentina", "Australia"),
  values = round(runif(9, 10, 25))
)

choropleth \%>\% 
  e_charts(countries) \%>\% 
  e_choropleth(values) \%>\% 
  e_visual_map(min = 10, max = 25)

choropleth \%>\% 
  e_charts(countries) \%>\% 
  e_map_3d(values, shading = "lambert") \%>\% 
  e_visual_map(min = 10, max = 30)
}

}
\seealso{
\code{\link{e_country_names}}
}
