% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{e_theme}
\alias{e_theme}
\alias{e_theme_custom}
\title{Themes}
\usage{
e_theme(e, theme)

e_theme_custom(e, theme)
}
\arguments{
\item{e}{An \code{echarts4r} object as returned by \code{\link{e_charts}}.}

\item{theme}{Theme, see below.}
}
\description{
Add a theme.
}
\section{Themes}{
 
\itemize{
  \item{\code{dark}}
  \item{\code{vintage}}
  \item{\code{westeros}}
  \item{\code{essos}}
  \item{\code{wonderland}}
  \item{\code{walden}}
  \item{\code{chalk}}
  \item{\code{infographic}}
  \item{\code{macarons}}
  \item{\code{roma}}
  \item{\code{shine}}
  \item{\code{purple-passion}}
  \item{\code{halloween}}
}
}

\examples{
mtcars \%>\% 
  e_charts(mpg) \%>\%
  e_line(disp) \%>\% 
  e_area(hp) \%>\% 
  e_x_axis(min = 10) -> p
  
p \%>\% e_theme("chalk")
p \%>\% e_theme_custom('{"color":["#ff715e","#ffaf51"]}')

}
\seealso{
\href{http://echarts.baidu.com/theme-builder/}{create your own theme}.
}
